speech-ai-pipeline/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP clients
httpx==0.25.2
aiofiles==23.2.1

# STT providers
openai==1.3.7
google-cloud-speech==2.21.0
azure-cognitiveservices-speech==1.34.0
speechrecognition==3.10.0

# LLM providers
anthropic==0.7.8
ollama==0.1.7

# TTS providers
google-cloud-texttospeech==2.16.3
elevenlabs==0.2.26
edge-tts==6.1.9
gTTS==2.4.0

# Audio processing
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1
numpy==1.24.3

# Utilities
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 